Separated animation control and implemented basic door animations
