+++ src/toniarts/openkeeper/world/animation/AnimationControl.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (C) 2014-2016 OpenKeeper
+ *
+ * OpenKeeper is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * OpenKeeper is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with OpenKeeper.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package toniarts.openkeeper.world.animation;
+
+import com.jme3.scene.control.Control;
+
+/**
+ * An interface for handling our animations, callbacks
+ *
+ * @author Toni Helenius <helenius.toni@gmail.com>
+ */
+public interface AnimationControl extends Control {
+
+    /**
+     * On animation stop
+     *
+     */
+    public void onAnimationStop();
+
+    /**
+     * Main animation cycle is done
+     */
+    public void onAnimationCycleDone();
+
+    /**
+     * Should we stop the animation
+     */
+    public boolean isStopAnimation();
+}
